Renamed service channel component
