+++ gutenberg/blocks/ts/src/service-location-service-channel-component.tsx
